requirements.txt
pytube~=12.1.0
tqdm~=4.64.0
pillow~=9.0.1
mutagen~=1.45.1